Functions fixed

+++ [9.1] Хэш-таблица/[9.1] Хэш-таблица/hashTable.h
@@ -1,13 +1,26 @@
 #pragma once
 #include "list.h"
 
+/* Описание структуры хеш-таблицы */
 struct HashTable;
 
+/* Создание хещ-таблицы */
 HashTable *createHashTable();
 
+/* Добавление слова в хеш-таблицу */
 void add(HashTable *hashTable, const std::string &word);
-float occupancyOfHashTable(HashTable *hashTable);
+
+/* Примерная заполняемость хеш-таблицы в процентах*/
+int occupancyOfHashTable(HashTable *hashTable);
+
+/* Максимальная длина листа */
 int maxLengthOfList(HashTable *hashTable);
-float averageLengthOfList(HashTable *hashTable);
+
+/* Средняя длина листа */
+int averageLengthOfList(HashTable *hashTable);
+
+/* Вывод содержимого хеш-таблицы на экран */
 void printHashTable(HashTable *hashTable);
+
+/* Удаление хеш-таблицы */
 void deleteHashTable(HashTable *hashTable);
